Update Range Maximum Query & Range Add Query.cpp

+++ Data Structure/Range Maximum Query & Range Add Query.cpp
@@ -1,8 +1,8 @@
 // lazy_segtree<S,op,e,F,mapping,composition,id> seg(n); で長さnのセグ木を構築。
 // ll ma=seg.prod(l,r); でa[l]~a[r-1]のmaxを求める。
 // seg.apply(l,r,x); でa[l]~a[r-1]にxを加える。
-using S=long long;
-using F=long long;
+using S=ll;
+using F=ll;
 const S INF=8e18;
 S op(S a,S b){return max(a,b);}
 S e(){return -INF;}
